Correct: C 1085 ~ 1090

+++ C/1085 [기초-종합] 소리 파일 저장용량 계산하기.c
@@ -6,7 +6,7 @@ int main(void) {
 	
 	scanf("%d %d %d %d", &hz, &bit, &channel, &sec);
 	
-	double size = (hz * bit * channel * sec);
+	double size = ((double)hz * (double)bit * (double)channel * (double)sec);
 	size /= (1024 * 1024 * 8);
 	
 	printf("%.1lf MB", size);
